pyproject.toml
[project]
name = "task1"
version = "0.1.0"
description = ""
authors = [
    {name = "RaiBack21",email = "sviatoslavrybak@icloud.com"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "alembic (>=1.15.2,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "cloudinary (>=1.44.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.test.dependencies]
sphinx = "^8.2.3"
pytest = "^8.3.5"
httpx = "^0.28.1"

[tool.pytest.ini_options]
pythonpath = ["."]